Ignore SIGPIPE and prefer to catch EPIPE and close everything
This is a well-known behavior. SIGPIPE appears sometimes when we try to
write something into a closed socket and, instead of to get a EPIPE, a
SIGPIPE is received by the system and the program finish. We prefer to
ignore SIGPIPE and correctly handle the EPIPE error by `write()`.

See this article about the behavior:
https://blog.nelhage.com/2010/02/a-very-subtle-bug/