Fix the test case and remove the default error handler
It still exists a case where we infinity-loop. It's when the error
handler responds. In that case, the client has a response and want to
continue but the protocol fallback to the client error handler too. We
finally want to retrieve the body of the response and the protocol wants
to finish the state machine. But we got the response before the error
handler so we don't cancel tasks... May be a fix is to separate the
Miou.Computation.t of the response and the error.